fix: 値段がdouble方化修正

+++ src/main/java/com/example/constants/Validate.java
@@ -4,7 +4,7 @@ package com.example.constants;
  * バリデーション用定数
  */
 public final class Validate {
-	public static final Integer PRICE_UPPER = 1000000000;
-	public static final Integer EXPENSE_PRICE_UPPER = 500000000;
+	public static final Double PRICE_UPPER = 1000000000.0;
+	public static final Double EXPENSE_PRICE_UPPER = 500000000.0;
 	public static final Integer TEXT_LENGTH = 2000;
 }
